requirements.txt
asgiref==3.7.2
azure-core==1.30.1
azure-storage-blob==12.19.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.9
django-oauth-toolkit==2.3.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
ecdsa==0.18.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
idna==3.6
isodate==0.6.1
jwcrypto==1.5.4
oauthlib==3.2.2
paho-mqtt==2.0.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tuya-connector-python==0.1.2
tuya-iot-py-sdk==0.6.6
typing_extensions==4.9.0
urllib3==2.2.1
websocket-client==1.7.0
whitenoise==6.6.0
